# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/storage/devices/snapshot-backup": {
            "post": {
                "description": "Aggregates all files from all managed devices onto the target snapshot-backup device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Start a snapshot backup to a device",
                "parameters": [
                    {
                        "description": "{targetDeviceSerial: string}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
        "/storage/devices/snapshot-backup/status/{jobId}": {
            "get": {
                "description": "Returns the current status of a snapshot backup job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Get snapshot backup job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
        "/storage/devices/snapshot-backup/verify": {
            "post": {
                "description": "Walks all files on the backup device and checks against the manifest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storage"
                ],
                "summary": "Verify integrity of a snapshot backup",
                "parameters": [
                    {
                        "description": "{deviceSerial: string, full: bool}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    }
}